move cli commands out of __init__ file
