+++ setup.py
@@ -77,7 +77,7 @@ setuptools.setup(
     python_requires=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
     entry_points={
         'console_scripts': [
-            'bentoml=bentoml.cli:cli',
+            'bentoml=bentoml.cli.commands:cli',
         ],
     },
     project_urls={
